+++ src/JSONSchema/index.ts
@@ -1,2 +1,2 @@
 export * from './types'
-export { getDataFromPath } from './path-handler'
+export { getDataFromPointer } from './path-handler'