+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `getDataFromPath` to `getDataFromPointer` as declared in the `README.md`.
+
 ## [0.2.0-beta.10] - 2020-02-21
 
 ### Fixed